memory.py
import google.generativeai as genai
from app_secrets import GEMINI_API_KEY
import time

genai.configure(api_key=GEMINI_API_KEY)

def load_history():
    """
    Simulates loading a massive context of medical history (5 years of PDFs).
    Uses Gemini's Long Context window.
    """
    print("Loading 5 years of medical history (simulated)...")
    
    # In a real scenario, we would upload 50+ PDF files using the File API
    # and wait for them to process.
    # history_files = [genai.upload_file(f) for f in glob.glob("history/*.pdf")]
    
    # Simulating a large context string for the prototype
    large_context_simulation = """
    PATIENT HISTORY SUMMARY DOES NOT REPLACE FULL RECORDS.
    
    2021-03-12: Vitamin D: 22 ng/mL (Low). Lipid Profile: Normal.
    2022-04-15: Vitamin D: 18 ng/mL (Low). Prescribed supplements.
    2023-01-20: Vitamin D: 25 ng/mL (Borderline).
    2024-05-10: Vitamin D: 30 ng/mL (Normal).
    2025-06-05: Vitamin D: 14 ng/mL (Critical Low). Patient reports fatigue.
    ... [Imagine 1 million tokens of detailed notes here] ...
    """
    
    model = genai.GenerativeModel("gemini-1.5-pro") # Supports 1M+ context
    
    prompt = """
    Review this entire patient history context. 
    Identify patterns in Vitamin D levels over the last 5 years.
    Are there seasonal trends or compliance issues?
    """
    
    print("Sending massive context to Gemini 3 Long Context window...")
    # Simulate processing delay
    time.sleep(1) 
    
    response = model.generate_content([large_context_simulation, prompt])
    return response.text

if __name__ == "__main__":
    print(load_history())
