requirements.txt
fastapi
uvicorn
google-generativeai
python-multipart
pydantic
requests
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
